+++ src/pages/admin/space/SpaceSearch.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function SpaceSearch(props) {
+	return <div></div>;
+}
+
+export default SpaceSearch;
